Create README.md
This project has been prepared for educational purposes. Time runs from the moment the user starts the exam. The wrong answer can be marked until the correct answer is found. In the exam, it is possible to switch between questions. Since the program is for educational purposes, it can answer the question once again when the previous questions are passed.